+++ livekit/sip.go
@@ -182,9 +182,25 @@ func (p *CreateSIPParticipantRequest) Validate() error {
 	}
 	if p.SipCallTo == "" {
 		return errors.New("missing sip callee number")
+	} else if strings.Contains(p.SipCallTo, "@") {
+		return errors.New("SipCallTo should be a phone number or SIP user, not a full SIP URI")
 	}
 	if p.RoomName == "" {
 		return errors.New("missing room name")
 	}
+	return nil
+}
+
+func (p *TransferSIPParticipantRequest) Validate() error {
+	if p.RoomName == "" {
+		return errors.New("missing room name")
+	}
+	if p.ParticipantIdentity == "" {
+		return errors.New("missing participant identity")
+	}
+	if p.TransferTo == "" {
+		return errors.New("missing transfer to")
+	}
+
 	return nil
 }